requirements.txt
numpy==2.3.2
paho-mqtt==2.1.0
pandas==2.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
schedule==1.2.2
six==1.17.0
tzdata==2025.2
